--- tomcat/src/main/java/com/controller/Controller.java
@@ -1,9 +1,0 @@
-package com.controller;
-
-import org.apache.coyote.http11.Http11Request;
-import org.apache.coyote.http11.Http11Response;
-
-public interface Controller {
-
-    Http11Response control(final Http11Request request);
-}
