+++ tomcat/src/main/java/org/apache/catalina/controller/RootController.java
@@ -1,7 +1,7 @@
-package com.controller;
+package org.apache.catalina.controller;
 
-import org.apache.coyote.http11.Http11Request;
-import org.apache.coyote.http11.Http11Response;
+import org.apache.coyote.http11.request.Http11Request;
+import org.apache.coyote.http11.response.Http11Response;
 
 public class RootController extends AbstractController {
