+++ tomcat/src/main/java/org/apache/catalina/controller/StaticResourcesController.java
@@ -1,13 +1,13 @@
-package com.controller;
+package org.apache.catalina.controller;
 
-import com.resolver.StaticResourceResolver;
-import com.resolver.StaticResourceResolver.ResolvedResource;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import org.apache.coyote.http11.Http11Request;
-import org.apache.coyote.http11.Http11Response;
+import org.apache.catalina.resolver.StaticResourceResolver;
+import org.apache.catalina.resolver.StaticResourceResolver.ResolvedResource;
+import org.apache.coyote.http11.request.Http11Request;
+import org.apache.coyote.http11.response.Http11Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
