+++ tomcat/src/main/java/org/apache/catalina/mapping/RequestKey.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.mapping;
 
 import java.util.Objects;
-import org.apache.coyote.http11.HttpMethod;
+import org.apache.coyote.http11.request.HttpMethod;
 
 public class RequestKey {
     private final String uri;
