+++ tomcat/src/main/java/org/apache/catalina/mapping/StaticControllerMapping.java
@@ -1,9 +1,9 @@
 package org.apache.catalina.mapping;
 
-import com.controller.Controller;
-import com.controller.StaticResourcesController;
 import java.util.ArrayList;
 import java.util.List;
+import org.apache.catalina.controller.Controller;
+import org.apache.catalina.controller.StaticResourcesController;
 
 public class StaticControllerMapping {
