+++ tomcat/src/main/java/org/apache/catalina/resolver/StaticResourceResolver.java
@@ -1,4 +1,4 @@
-package com.resolver;
+package org.apache.catalina.resolver;
 
 import java.net.URL;
