+++ tomcat/src/main/java/org/apache/coyote/http11/request/Http11RequestHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.ArrayList;
 import java.util.List;
