+++ tomcat/src/main/java/org/apache/coyote/http11/response/Http11Response.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 import java.nio.charset.StandardCharsets;
 
@@ -17,28 +17,28 @@ public class Http11Response {
     }
 
     public static Http11Response ok(final String contentType, final String body) {
-        Http11ResponseHeaders headers = new Http11ResponseHeaders();
+        Http11ResponseHeaders headers = Http11ResponseHeaders.create();
         headers.addHeader("Content-Type", contentType);
         headers.addHeader("Content-Length", String.valueOf(body.getBytes(StandardCharsets.UTF_8).length));
         return new Http11Response(HttpStatus.OK, headers, body);
     }
 
     public static Http11Response notFound(final String contentType, final String body) {
-        Http11ResponseHeaders headers = new Http11ResponseHeaders();
+        Http11ResponseHeaders headers = Http11ResponseHeaders.create();
         headers.addHeader("Content-Type", contentType);
         headers.addHeader("Content-Length", String.valueOf(body.getBytes(StandardCharsets.UTF_8).length));
         return new Http11Response(HttpStatus.NOT_FOUND, headers, body);
     }
 
     public static Http11Response redirect(final String location) {
-        Http11ResponseHeaders headers = new Http11ResponseHeaders();
+        Http11ResponseHeaders headers = Http11ResponseHeaders.create();
         headers.addHeader("Location", location);
         headers.addHeader("Content-Length", "0");
         return new Http11Response(HttpStatus.FOUND, headers, "");
     }
 
     public static Http11Response redirect(final String location, final String cookieHeader) {
-        Http11ResponseHeaders headers = new Http11ResponseHeaders();
+        Http11ResponseHeaders headers = Http11ResponseHeaders.create();
         headers.addHeader("Location", location);
         headers.addHeader("Set-Cookie", cookieHeader);
         headers.addHeader("Content-Length", "0");
@@ -47,7 +47,7 @@ public class Http11Response {
 
     public static Http11Response serverError() {
         String body = "Internal Server Error";
-        Http11ResponseHeaders headers = new Http11ResponseHeaders();
+        Http11ResponseHeaders headers = Http11ResponseHeaders.create();
         headers.addHeader("Content-Type", "text/html;charset=utf-8");
         headers.addHeader("Content-Length", String.valueOf(body.getBytes(StandardCharsets.UTF_8).length));
         return new Http11Response(HttpStatus.INTERNAL_SERVER_ERROR, headers, body);
