+++ tomcat/src/main/java/org/apache/coyote/http11/response/Http11ResponseHeaders.java
@@ -1,12 +1,19 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class Http11ResponseHeaders {
+class Http11ResponseHeaders {
     private static final String CRLF = "\r\n";
 
     private final Map<String, String> headers = new LinkedHashMap<>();
+
+    private Http11ResponseHeaders() {
+    }
+
+    static Http11ResponseHeaders create() {
+        return new Http11ResponseHeaders();
+    }
 
     public void addHeader(final String name, final String value) {
         headers.put(name, value);
