+++ tomcat/src/test/java/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import org.apache.catalina.FrontController;
+import org.apache.coyote.http11.processor.Http11Processor;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
